+++ product-service/src/main/java/com/example/productservice/services/ProductServiceImpl.java
@@ -54,8 +54,8 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public int countProducts() {
-        return (int) productRepository.count();
+    public Long countProducts() {
+        return productRepository.count();
     }
 
     @Override
